# api-gateway/src/swagger/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "ZyvoPay API Gateway",
    "description": "ZyvoPay Payment Gateway API - Complete banking-as-a-service platform for PIX payments, TED transfers, and financial operations.\n\n## Authentication\nAll endpoints (except `/health` and `/api/auth/clients`) require API Key authentication.\n\nProvide your API key via:\n- `X-API-Key` header, or\n- `Authorization: Bearer <api_key>` header\n\n## Rate Limiting\n- Default: 60 requests per minute per API key\n- Rate limit headers are returned in responses\n\n## Webhooks\nConfigure your webhook URL in the client settings to receive real-time event notifications.\n",
    "version": "1.0.0",
    "contact": {
      "name": "ZyvoPay Support",
      "email": "support@zyvopay.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Development server"
    },
    {
      "url": "https://api.zyvopay.com",
      "description": "Production server"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      },
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
...
  },
  "security": [
    {
      "ApiKeyAuth": []
    },
    {
      "BearerAuth": []
    }
  ],
  "paths": {
...
    "/api/transactions": {
      "get": {
        "summary": "List transactions",
        "tags": [
          "Transactions"
        ],
        "parameters": [
          {
            "name": "subaccount_id",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "transaction_type",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "start_date",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "end_date",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 50
            }
          },
          {
            "name": "offset",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of transactions"
          }
        }
      }
    },
    "/api/transactions/{id}": {
      "get": {
        "summary": "Get transaction details",
        "tags": [
          "Transactions"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Transaction details"
          }
        }
      }
    },
    "/api/transactions/statement/{subaccount_id}": {
      "get": {
        "summary": "Get account statement",
        "tags": [
          "Transactions"
        ],
        "parameters": [
          {
            "name": "subaccount_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "start_date",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "end_date",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Account statement"
          }
        }
      }
    },
...
  }
}